Improved code structure and operand parsing
All operands as well as the code interpreter both contain references to a simulation object that holds all the data, which makes it easier for the operands to execute branch-type opcodes.
In addition, all operations can now determine the type of their operands as well as parse it into an integer (for now only immediate addressing type).